Waited for the server navigation to render
When there's a server navigation the deserialize resolves straight away because it returns null instead of rsc content. That would mean a blank screen, for the original rsc fetch navigation, waiting for the server navigation content to load.
So waited for the server navigation during the original rsc fetch so that react doesn't commit until the server navigation content is loaded - then when the client redirect happens the new content is already there. Waited for it with use - had to get at the underlying internal promise because user doesn't except the object. Is there another way so don't have to get at react internals?