Forced the root to refetch when redirect on server
Scenario is search People and redirect to People on server (maybe change page or sort, for example) - the root people scene view has refetch = [] so it won't refetch and then won't get the rsc content waiting for it (the list will get it because it always refetches and then loading people content inside list goes very wrong - infinite nesting?!).
When redirect need to force the root scene view to refetch so it gets the rsc content already cached. Used ignoreCache to force the refetch (need to check for repercussions).
It still wasn't working, but turned out was using the wrong navigation event in the deserialize - the deferred lags behind so the cache check was failing in this refresh scenario. Need to use the navigation event because that's got the latest value and the latest value is used in the fetch/deserialize call.
This problem only happens with refresh redirect because other navigations use startTransition so the deferred event always matches the navigation event.