Detected redirect based on oldState
There's only oldState on the server if there's been a navigation on the server. Also only sent url back to client if there's been a server navigation - simpler than working it out on the client (probably have to compare all crumb states and data because query param order could be different so can't just compare urls).
Also renamed things for clarity - for example, rscNavigation/Cache seems better than Navigate because it's happened and the cache matches with SceneRscView