Made refetch scene view work with rsc navigate
With refetch(), used to change the navigation event just for that scene view. But that doesn't work with rsc navigate/redirect because that functions at the navigation handler level. So changed the navigation event at the navigation handler level instead - used the ignoreCache to store the key of the scene view that requests the refetch.
This actually fixes a bug with the previous implementation - if the ancestor was marked as fetching from previous navigation then the refetch wouldn't fire anyway. So changing the navigation event just for the scene view was already flawed - just more obvious with rsc navigation.
In testing, created a Test component as child of List that triggers a refetch in timeout on first load of people scene. This didn't work because there was no old state so the people scene view was marked as fetching - this ancestor fetching prevented the list refetch. So added check that ignores the old getShow if ignoring cache