Prevented refetch picking up old rsc nav/redirect
Created a Test component that's a child of list scene view and it refetches list after 5 seconds. Then load people and do a search, for example, that does a rsc navigate/redirect. Then after 5 seconds the list refetches but it doesn't go to the server because it picked up the rsc navigation/redirect when it called deserialize - the state context hasn't changed from the redirect (just has ignoreCache in data but the state context is the same) so it gets the redirect cache - this makes the page hang because it returns cache people response for the list refetch request (loading people inside people).
So added a new indicator on navigation event data that says if it's a rsc redirect navigation event. Then only return the cached response when that's true. Make sure to reset it when doing a refetch request