Prevented caching history when rsc navigation
Scenario is navigation to A then B then to A but rsc navigate to C. Then if press browser back twice it will go back to A and show a blank page. The 'blank' response from the second A is added to history cache - rsc navigation gives blank response because the response is for scene C (not A). So when browser back twice from C, the cached blank scene is returned.
Only need to browser back twice because rsc navigation isn't added to browser history. Changed to do the same for the browser history cache and not cache the result from rsc navigation. Not on the original navigation event, anyway - it will be cached once the navigation happens against the correct new navigation event.
Stored the original navigation event in the history cache when there's rsc navigation so that the scene view knows not to write to history cache. Interestingly, it is ok to update the ref during this background render (although can't use refs as the rsc cache for scene views - did try this when first doing rsc).
In scenario where navigation from A to B and rsc navigate to C, then history cache is created for B in order to store the original navigation event. But there's no count against the cache, so changed the history delete to check when there's no count (gap is NaN) and delete the entry