Refetched on hmr reload instead of setRoot
Hmr threw some error about registerSceneView - didn't bother investigating because want to move away from setRoot. It doesn't work on mobile because it loses the React state of the crumbs in the stack. Prefer to refetch - there are issues with crumbs, here, too, because will only refetch the current scene, not the crumbs. But prefer to use refetch (and can maybe refetch crumbs on navigation in the future).
Refetching for hmr works apart from when modifying stateNavigator route for the current scene. Will see if can fix that but still ok if can't