Handled hmr when current route changed
When edit route in state navigator for current scene then get client side error saying url invalid. So copied the logic that used to be in put action on the server and moved into catch block on the client. Stored the old state navigator so can build the new url from the old state context - the one before the hmr.
Not sure why the deserialize/refetch on the old state navigator (onHmrReload) doesn't happen - guessing react ignores it because the navigation provider updates which makes it redundant? anyway, it's good that don't have to handle a 'invalid url' on the server side because it never reaches