Supported rsc navigate to multiple roots
Can have more than one scene view so need to return them all, not just the first/last one. Returned object with key of the scene view name and value the render content. Changed navigation handler to expect sceneViews instead of sceneView. The use trick to await the render isn't right because it's awaiting them in series - took 5 + 2 seconds = 7 instead of expected 5?! Not sure if this is a use trick problem?!