Corrected to match the deserialize
Should be using the navigation event data because the deferred lags behind. It's still not pausing for the refresh case though - there's something wrong with the waiting when it comes to people. Think it's because there's no async data fetch in people at the top so the promise returns something and the use completes? but the render hasn't finished so it shouldn't?!