Prevented white screen for rsc navigate
Scenario is added delay to people fetch then when navigating back from People to Person there was a blank screen. The use trick in NavigationHandler was meant to wait for the content to render but it doesn't work - that only waits for the promise to resolve, and in the case of People, it resolves instantly. That's because there's no data fetch in the root scene view - the fetch is lower down in the list. So the root scene view resolves, the use trick completes and there's a blank screen - the blank screen is because when react renders the rsc navigate result - the people scene - it then waits for the list to complete - there's no suspense boundary, so it won't show the root scene view until the list resolves.
Similar problem, but not exactly the same, if rsc refresh - so  click a sort link to go from People to People then see a half blank screen. The reason it's half blank, not completely blank, is because the list is deferred but the root scene view isn't (refresh) - actually not sure about that because the rsc navigate on the client is wrapped in startTransition - even for refresh - so not sure why?!
Anyway, the only way to avoid blank screen is to render the rsc navigated content - that way react decides when things are complete. So overrode the show of rsc scene view to see if it's in a rsc navigate scenario - if it is then show the rsc navigated to content. So ignore the show because that's from the original navigation event and need to show the data from the new rsc navigate event - used the history cache to determine if rsc navigate and then took the content from there. If there's no content then return children because there will only be root content and want the children to also return their content