Ensured rsc navigate has the right state context
Got round the white screen in #d56f5ad by rendering the rsc navigate content - rendering is the only way to ensure React waits for the content before showing (avoiding the white screen). The problem was that was showing the rsc navigate content with the original state context - so there was a mismatch.
Take for example, navigate from People to People with refresh by clicking page 2, for example) by rsc navigate to Person. This would show the person but against the people page 2 state context - then it would navigate immediately to the person to correct things. But there's a brief time where the context and content don't match. This is no good - someone could have a client component and do something in useEffect and assume state is person, but it won't be and so would go wrong.
Have to make the state context and content match.
Realised there's a much better way of handling the rsc navigation by using the state context as the 'cache'!! This is available to scene views as well so don't need to use history to communicate between handler and scene view. So just store the rsc navigate response on the state context. Then also mutate the state context to match the rsc navigation!! Not sure why didn't do this before - though not allowed to mutate - but was already mutating the ref so this is no different. Mutate state context rather than overwrite because want to keep the same object reference - important part of navigation router is that 1 navigation means === 1 state context.
Then the state context matches the content and don't need the weird scene view logic that returned rsc navigate content even when it wasn't showing - now it will only return content if it's showing because context matches content.