Fixed refetch after rsc navigate
When refetching the state context stays the same, intentionally (should only change when navigating). But that meant the rsc flag was still there and that prevented the refetch - fetching bails early if it sees rsc flag because it knows the content is already fetched during rsc navigate.
Changed to move the rsc flag from state context to the navigation event - that way can clear it when mutating the navigation event for refetch. So raiseNavigationEvent checks if there's rsc flag on state context and moves it.
Then updated scene view to check both places (state context and navigation event) to see if it's rsc navigation.