Avoid redundant follow up rsc navigation
When rsc navigation there was 2 navigations/setState on the client - the first caused by the initial navigation and then the second one is the immediate follow up in useEffect that navigates the right place (the rsc navigation result).
But this was legacy of how used to do it, when the first setState had the wrong context - the original state context. So needed the follow up to correct it. But not doing it like that anymore - correcting/mutating the state context immediately when getting the rsc response back. So there's no need for the follow up navigation.
This is simpler to follow and also means don't have to worry about what happens if navigating happens while it was on the interim navigation/state - there's a slim chance that a navigation could get in before the follow up. But now the first navigation/state is correct there's no worries.
Still need to copy the rsc from state context to navigation context so can clear it when mutating the navigation event for refetch, for example.