Forced render after navigation event mutated
Scenario is People then sort then browser back then sort - would get blank list - similar problem on Person then friends then browser back then friends gives blank friends.
Not sure why but after a browser back React doesn't do a render after the navigation event is mutated. So the render is with the unmutated/original navigation event and list doesn't return children from scene view - it doesn't think ancestor is fetching because there's been no extra render with the mutated/latest navigation event.  So it's left with the original cache promise which returns null because of the rsc navigate.
Not sure why this only goes wrong with browser history but doing the same test with navigation (or setting cachedHistory to false) doesn't give the problem. It's a timing bug so believable the cached history makes it take a different path?
Adding a delay, well, not resolving the promise immediately causes that extra render which gives the correct content