Tried to add rsc navigate to mobile
The scene views on mobile aren't rendered until the scene is added to the stack. So can't redirect to them because they haven't been registered as root views. Exposed the registry as a hook and called it from the navigation stack for all scenes.
This allowed the rsc navigate but it still doesn't work. Not sure why but React doesn't do another render after mutating the navigation event from the rsc navigate result. So renders with the old navigation event instead of the new. But even if it did do a render it still wouldn't work because the Navigation Stack doesn't update its keys (motion state) unless the navigation event changes - so mutating it is no good.