Navigated again instead of mutating
No idea why didn't realise this worked sooner - got stuck on thinking couldn't update state in the background render. But seen loads of examples that show just that - as long as the state update is wrapped in a transition/deferred then it works. So removed all the mutation and did another navigation based on the rsc navigate url. Made sure it's done against the real state navigator context so that the original navigation doesn't appear in old state/data on the rsc navigate context.
This fixes mobile and is obviously the right way to do it.