Prevented resolution to avoid blank screen
Scenario is People then sort and rsc navigate to person then browser back and sort again for rsc navigate to person - then get a blank list on people briefly before the the person shows. This only happens when using the history cache - not show why it's different but it is. Obviously the promise is resolving and then the people scene updates with null for list (because rsc navigate doesn't return list) before the rsc navigate happens. Prevented the promise from resolving to stop this - there's no need for this to resolve because it's the next navigation event that has to resolve - the rsc navigate one. The first navigation event isn't for committing so no need to resolve.
Forgot to explain when first committed so undid it and did it again so could add explanation.