Fixed hmr navigate then update route
Scenario is navigate from People to Person then update the person route. This navigated back to People because it ran NavigationProvider from scratch which took the original url prop - pointing at People.
Changed to use the current url cached on the oldStateNavigator. But this by itself gave a blank screen because the Person scene view doesn't refetch but there's nothing in the rsc cache. It doesn't refetch because the data hasn't changed (same url) and there's nothing in the rsc cache because there's a new navigation event for the new state navigator.
Could copy across the rsc cache from old navigation event to the new one when the state navigator changes. That would fix the blank screen - but wouldn't fix changing other state info. For example, changing default show from false to true would expand but show blank friends - because there's no refetch and there's no friends in children or rsc cache.
Instead changed to set ignoreCache to true (pass {} as rsc cache) so that always refetches when state navigator changes.
Also had to rejig the oldStateNavigator because using window in the useMemo didn't play well - parcel kept giving weird errors on load (TypeError: Failed to execute 'getReader' on 'ReadableStream': ReadableStreamDefaultReader constructor can only accept readable streams that are not yet locked to a reader).