Copied over hmr changes to mobile
Hmr bug on mobile - navigate from People to Person then edit person route then expand friends then click link back to People and it would fall over. There was still a history listener for the oldStateNavigator hanging around and it threw because it had the old route.
The historyManager at the top wasn't global so it was a different one when the hmr ran the NavigationProvider from scratch. So got the right one from the oldStateNavigator and stopped it