Copied over rsc navigate get request from web
Didn't copy over the replace history from NavigationProvider because want the history back people if load person scene. Replace history works on web where don't want the history but on mobile do want it so add is better (doesn't really matter that it will do add for hmr as well if route changes, for example, because there will be other history there that can't change anyway - could check for oldStateNavigator and do replace but it's ok as is for now).