Primed the navigation on the server instead
So if load the Person scene then need to prime state context with People scene so that it loads with the correct stack (and browser history padded with people scene). Prefer to do this on the server now it's possible - makes the NavigationProviders on web and mobile almost matching. It was happening on the server before as well, but this is in the express handler and better that it has the correct stack now rsc navigate can happen there too