Removed root property for consistency
Didn't realise before that could rsc serialize any object so got confused about root - and even more confused about the await on the client. So removed root and await for consistency. Will bring in object properties when copying over the rsc navigate from parcel sample but want a clean baseline to start with