Copied over hmr changes from parcel
Hmr works except when changing the current route - the hmr update fires before the latest code is loaded. There was [hmr bug around this anyway](https://github.com/vitejs/vite-plugin-react/issues/666) - although before a try/catch allowed it to work, this time get a blank screen. The error this time is thrown on the server and before it was on the client - that explains why blank screen this time and try/catch worked before