Used Activity instead of Suspense if available
Using Suspense to freeze crumb scenes is a hack and it doesn't work when using RSC data fetching. React looks for Suspended components so it knows when to commit and show fallback - so had to turn off Freeze when using RSC. Replacing Suspense with Activity means can turn Freeze back on for RSC - done basic test of logging every second in People scene and there's no log when on the Person scene.