Reenabled shared el by forcing layout
Scenario is a navigation stack per primary tab with shared element in one of the tab stacks. Changing tab prevents the shared element from working - so shared element would only work on app load and before tab change.
The onStart of DefaultSpecialEffectsController checks if the container (NavigationStackView) has been laid out and, if not, skips the transitions. The isLaidOut of View returns false if layout hasn't been called since last added to window. Added the standard requestLayout code for react native views so that layout is called when the tab changes and the NavigationStackView is added to the window.
Thanks to @afkcodes for letting me know about the problem over twitter dm. Pushing so that he can try it out in his code